requirements.txt
Flask==3.0.0
celery==5.3.6
redis==5.0.1
boto3==1.34.0
numpy==1.26.2
scikit-learn==1.3.2
scikit-image==0.22.0
Pillow==10.1.0
gunicorn==21.2.0